created TopicsList component
