+++ Client/src/pages/Repartition.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import NavBar from '../components/NavBar'
 import LeftContainer from '../container/LeftContainer'
-
+import TopicsList from '../components/PartitionComponents/TopicsList'
 
 
 // creating repartition page that has navbar and the repartition container that will have:
@@ -13,14 +13,15 @@ import LeftContainer from '../container/LeftContainer'
 function Repartition() {
   return (
     <>
-        <div className='root'>
-            <NavBar />
-            <div className='homeContainer'>
-                <LeftContainer />
-                <div className='rightContainer' style={{ minWidth: '400px' }}>
-                </div>
-            </div>
+      <div className='root'>
+        <NavBar />
+        <div className='homeContainer'>
+          <LeftContainer />
+          <div className='rightContainer' style={{ minWidth: '400px' }}>
+            <TopicsList />
+          </div>
         </div>
+      </div>
     </>
   )
 }
